Убран сервис почты с Gmail. Создан новый Yandex. Настройки соотвественно тоже изменены.
