+++ Core/OnlineSchool.App/Course/Commands/CreateCourse/CreateCourseCommandHandler.cs
@@ -1,8 +1,11 @@
 ﻿using ErrorOr;
 using MediatR;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Configuration;
 using OnlineSchool.App.Common.Interfaces.Persistence;
 using OnlineSchool.App.Common.Interfaces.Services;
 using OnlineSchool.Domain.Course;
+using static OnlineSchool.Domain.Common.Errors.Errors;
 
 namespace OnlineSchool.App.Course.Commands.CreateCourse;
 
@@ -10,10 +13,14 @@ public class CreateCourseCommandHandler
     : IRequestHandler<CreateCourseCommand, ErrorOr<bool>>
 {
     private readonly IUnitOfWork _unitOfWork;
+    private readonly IEmailService _emailService;
+    private readonly ILogger<CreateCourseCommandHandler> _logger;
 
-    public CreateCourseCommandHandler(IUnitOfWork unitOfWork)
+    public CreateCourseCommandHandler(IUnitOfWork unitOfWork, IEmailService emailService, ILogger<CreateCourseCommandHandler> logger)
     {
         _unitOfWork = unitOfWork; 
+        _emailService = emailService;
+        _logger = logger;
     }
 
     public async Task<ErrorOr<bool>> Handle(
@@ -25,8 +32,8 @@ public class CreateCourseCommandHandler
 
         //2. Добавляем курс в хранилище
         if (await _unitOfWork.Courses.Add(course))
-            return await _unitOfWork.CompleteAsync();
+			return await _unitOfWork.CompleteAsync();
 
-        return false;
+		return false;
     }
 }
