+++ Infrastructure/OnlineSchool.Infrastructure/Services/Email/EmailYandexSettings.cs
@@ -6,12 +6,12 @@ using System.Threading.Tasks;
 
 namespace OnlineSchool.Infrastructure.Services.Email
 {
-	public class EmailGoogleSettings
+	public class EmailYandexSettings
 	{
 		public const string SectionName = "EmailSettings";
 
 		public string Address { get; init; } = null!;
-		public string GmailPassword { get; init; } = null!;
+		public string YandexPassword { get; init; } = null!;
 		public string From { get; init; } = null!;
 
 	}
