+++ Presentation/OnlineSchool.API/appsettings.json
@@ -9,12 +9,12 @@
     "API_KEY": "AIzaSyBa0PWVV4-_zKEnIiXPsDXVyH4gNRjmrzA"
   },
   "EmailSettings": {
-    "Address": "onlineschoolproject2023@gmail.com",
-    "GmailPassword": "supyootzbubmagno",
-    "From": "OnlineSchool"
+    "Address": "OnlineSchool2023@yandex.ru",
+    "YandexPassword": "edqajhqtshxfbewr", 
+    "From": "По Шагам"
   },
   "ConnectionStrings": {
-    "SqlServer": "Data Source=LAPTOP-IGE01LPP\\SQLEXPRESS;Initial Catalog=OnlineSchoolDb;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
+    "SqlServer": "data source=VLAD\\VLAD;initial catalog=OnlineSchoolDb;trusted_connection=true"
     //@"data source=LAPTOP-9S2AK2B9;initial catalog=OnlineSchoolDB;trusted_connection=true"
     //"Data Source=LAPTOP-IGE01LPP\\SQLEXPRESS;Initial Catalog=OnlineSchoolDb;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
   },
